refactor: 기존의 doamin,enums 패키지, VideoRepository 다시 작성
