+++ dev-route/src/main/java/com/teamdevroute/devroute/video/VideoRepository.java
@@ -1,0 +1,7 @@
+package com.teamdevroute.devroute.video;
+
+import com.teamdevroute.devroute.video.domain.Videos;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface VideoRepository extends JpaRepository<Videos,Long> {
+}
