+++ dev-route/src/main/java/com/teamdevroute/devroute/video/enums/PlatformName.java
@@ -1,0 +1,5 @@
+package com.teamdevroute.devroute.video.enums;
+
+public enum PlatformName {
+    Youtube,Infrean,Udemy
+}
