+++ dev-route/src/main/java/com/teamdevroute/devroute/video/enums/TechnologyStackName.java
@@ -1,0 +1,5 @@
+package com.teamdevroute.devroute.video.enums;
+
+public enum TechnologyStackName {
+    htmlcss,파이썬,자바,자바스크립트,스프링,Nodejs,React,Vuejs,Angular,
+}
